Holy updates of updates
added the following:
Map for class related objects.

Map for the pictures where we can take them out of.

Adding the Page: MemoryGrid.
which makes the grid needed to play the game.

Adding the Page: MemoryPage.
which has the following in it already:
playable game, Highscore properities, and a subway to save.