The next one
the following update adds the following:
bug fixes in the pages:
memorygrid, Welkompage, MemoryPage.
added 2 classes for the datagridview for the upcomming highscore.
added a way to create a save state for memory games called ObjectSterializer.

Through the help of Arnold Verruit added some Backgroundimages for the back sides of cards.

Put them in the Resource/Themes/ appropriate theme

Co-Authored-By: Rambokes <55061625+Rambokes@users.noreply.github.com>